get_data/moneypuck/pyproject.toml
[project]
name = "moneypuck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "boto3>=1.38.13",
    "dbt-duckdb>=1.9.3",
    "duckdb>=1.2.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.9",
]
